+++ tests/mocks.test.ts
@@ -165,6 +165,14 @@ export const mockBrowser = {
 
 	runtime: {
 		lastError: undefined as any,
+		_mockUninstallUrl: "",
+		setUninstallURL(url: string, callback?: () => void): void {
+			this._mockUninstallUrl = url;
+			callback?.();
+		},
+		getMockUninstallURL(): string {
+			return this._mockUninstallUrl;
+		},
 		sendMessage: (
 			message: InternalMessage,
 			callback?: (response?: any) => void,
@@ -598,6 +606,9 @@ export const translations: Record<
 		},
 		"plz_use_sf_login": {
 			"message": "$you_should $use_sf_login please!",
+		},
+		"prompt_with_newline": {
+			"message": "Enjoying $sf_login?\nPlease continue.",
 		},
 		"error_missing_key": {
 			"message": "Key not found anywhere",